bot/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
firebase-admin==6.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-telegram-bot==20.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
grpcio==1.58.0
gunicorn==21.2.0
uvicorn==0.25.0
fastapi==0.108.0
google-auth==2.23.0
google-api-core==2.11.1
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
googleapis-common-protos==1.60.0
yookassa==3.0.0
pytz==2023.3.post1
